#527 collection_name will have its commas removed if used for arr tags
